refactor: extract ComponentDemo to a separate file for reusability
